Replace fakes with mocks
