+++ core/control-plane/contract/src/test/java/org/eclipse/dataspaceconnector/contract/negotiation/ContractNegotiationIntegrationTest.java
@@ -15,60 +15,144 @@
 
 package org.eclipse.dataspaceconnector.contract.negotiation;
 
+import org.eclipse.dataspaceconnector.common.statemachine.retry.SendRetryManager;
 import org.eclipse.dataspaceconnector.common.util.junit.annotations.ComponentTest;
-import org.eclipse.dataspaceconnector.spi.contract.negotiation.observe.ContractNegotiationListener;
+import org.eclipse.dataspaceconnector.contract.observe.ContractNegotiationObservableImpl;
+import org.eclipse.dataspaceconnector.core.controlplane.defaults.negotiationstore.InMemoryContractNegotiationStore;
+import org.eclipse.dataspaceconnector.policy.model.Action;
+import org.eclipse.dataspaceconnector.policy.model.Duty;
+import org.eclipse.dataspaceconnector.policy.model.Policy;
+import org.eclipse.dataspaceconnector.policy.model.PolicyType;
+import org.eclipse.dataspaceconnector.spi.command.CommandQueue;
+import org.eclipse.dataspaceconnector.spi.command.CommandRunner;
+import org.eclipse.dataspaceconnector.spi.contract.ContractId;
+import org.eclipse.dataspaceconnector.spi.contract.validation.ContractValidationService;
+import org.eclipse.dataspaceconnector.spi.entity.StatefulEntity;
+import org.eclipse.dataspaceconnector.spi.iam.ClaimToken;
+import org.eclipse.dataspaceconnector.spi.message.RemoteMessageDispatcherRegistry;
+import org.eclipse.dataspaceconnector.spi.monitor.ConsoleMonitor;
+import org.eclipse.dataspaceconnector.spi.policy.store.PolicyDefinitionStore;
+import org.eclipse.dataspaceconnector.spi.response.StatusResult;
 import org.eclipse.dataspaceconnector.spi.result.Result;
+import org.eclipse.dataspaceconnector.spi.types.domain.asset.Asset;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.agreement.ContractAgreement;
+import org.eclipse.dataspaceconnector.spi.types.domain.contract.agreement.ContractAgreementRequest;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.ContractNegotiation;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.ContractOfferRequest;
+import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.ContractRejection;
+import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.command.ContractNegotiationCommand;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.offer.ContractOffer;
+import org.jetbrains.annotations.NotNull;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.mockito.stubbing.Answer;
 
+import java.net.URI;
 import java.time.Duration;
+import java.time.ZonedDateTime;
+import java.util.ArrayList;
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
 
+import static java.util.concurrent.CompletableFuture.completedFuture;
+import static java.util.concurrent.CompletableFuture.failedFuture;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 import static org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.ContractNegotiationStates.CONFIRMED;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.ArgumentMatchers.isA;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ComponentTest
-class ContractNegotiationIntegrationTest extends AbstractContractNegotiationIntegrationTest {
+class ContractNegotiationIntegrationTest {
 
     private static final Duration DEFAULT_TEST_TIMEOUT = Duration.ofSeconds(15);
     private static final Duration DEFAULT_POLL_INTERVAL = Duration.ofMillis(100);
-    private final ContractNegotiationListener negotiationListenerMock = mock(ContractNegotiationListener.class);
+    private final InMemoryContractNegotiationStore providerStore = new InMemoryContractNegotiationStore();
+    private final InMemoryContractNegotiationStore consumerStore = new InMemoryContractNegotiationStore();
+    private final ContractValidationService validationService = mock(ContractValidationService.class);
+    private final RemoteMessageDispatcherRegistry providerDispatcherRegistry = mock(RemoteMessageDispatcherRegistry.class);
+    private final RemoteMessageDispatcherRegistry consumerDispatcherRegistry = mock(RemoteMessageDispatcherRegistry.class);
+    protected ClaimToken token = ClaimToken.Builder.newInstance().build();
+    private String consumerNegotiationId;
+
+    private ProviderContractNegotiationManagerImpl providerManager;
+    private ConsumerContractNegotiationManagerImpl consumerManager;
+
+    @BeforeEach
+    void init() {
+        var monitor = new ConsoleMonitor();
+
+        CommandQueue<ContractNegotiationCommand> queue = (CommandQueue<ContractNegotiationCommand>) mock(CommandQueue.class);
+        when(queue.dequeue(anyInt())).thenReturn(new ArrayList<>());
+
+        CommandRunner<ContractNegotiationCommand> runner = (CommandRunner<ContractNegotiationCommand>) mock(CommandRunner.class);
+
+        SendRetryManager<StatefulEntity> sendRetryManager = mock(SendRetryManager.class);
+
+        providerManager = ProviderContractNegotiationManagerImpl.Builder.newInstance()
+                .dispatcherRegistry(providerDispatcherRegistry)
+                .monitor(monitor)
+                .validationService(validationService)
+                .waitStrategy(() -> 1000)
+                .commandQueue(queue)
+                .commandRunner(runner)
+                .observable(new ContractNegotiationObservableImpl())
+                .store(providerStore)
+                .policyStore(mock(PolicyDefinitionStore.class))
+                .sendRetryManager(sendRetryManager)
+                .build();
+
+        consumerManager = ConsumerContractNegotiationManagerImpl.Builder.newInstance()
+                .dispatcherRegistry(consumerDispatcherRegistry)
+                .monitor(monitor)
+                .validationService(validationService)
+                .waitStrategy(() -> 1000)
+                .commandQueue(queue)
+                .commandRunner(runner)
+                .observable(new ContractNegotiationObservableImpl())
+                .store(consumerStore)
+                .policyStore(mock(PolicyDefinitionStore.class))
+                .sendRetryManager(sendRetryManager)
+                .build();
+    }
+
+    @AfterEach
+    void tearDown() {
+        providerManager.stop();
+        consumerManager.stop();
+    }
 
     @Test
     void testNegotiation_initialOfferAccepted() {
+        when(providerDispatcherRegistry.send(any(), isA(ContractAgreementRequest.class), any())).then(onProviderSentAgreementRequest());
+        when(consumerDispatcherRegistry.send(any(), isA(ContractOfferRequest.class), any())).then(onConsumerSentOfferRequest());
         consumerNegotiationId = "consumerNegotiationId";
         ContractOffer offer = getContractOffer();
         when(validationService.validate(token, offer)).thenReturn(Result.success(offer));
         when(validationService.validate(eq(token), any(ContractAgreement.class),
                 any(ContractOffer.class))).thenReturn(true);
 
-        // Create and register listeners for provider and consumer
-        providerObservable.registerListener(negotiationListenerMock);
-        consumerObservable.registerListener(negotiationListenerMock);
-
         // Start provider and consumer negotiation managers
         providerManager.start();
         consumerManager.start();
 
         // Create an initial request and trigger consumer manager
-        ContractOfferRequest request = ContractOfferRequest.Builder.newInstance()
+        var request = ContractOfferRequest.Builder.newInstance()
                 .connectorId("connectorId")
                 .connectorAddress("connectorAddress")
                 .contractOffer(offer)
                 .protocol("protocol")
                 .build();
-        consumerManager.initiate(request);
 
+        consumerManager.initiate(request);
 
         await().atMost(DEFAULT_TEST_TIMEOUT)
                 .pollInterval(DEFAULT_POLL_INTERVAL)
@@ -87,31 +171,22 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
                     verify(validationService, atLeastOnce()).validate(token, offer);
                     verify(validationService, atLeastOnce()).validate(eq(token), any(ContractAgreement.class), any(ContractOffer.class));
                 });
-
-
-        // Stop provider and consumer negotiation managers
-        providerManager.stop();
-        consumerManager.stop();
     }
 
     @Test
     void testNegotiation_initialOfferDeclined() {
+        when(consumerDispatcherRegistry.send(any(), isA(ContractOfferRequest.class), any())).then(onConsumerSentOfferRequest());
         consumerNegotiationId = null;
         ContractOffer offer = getContractOffer();
 
         when(validationService.validate(token, offer)).thenReturn(Result.success(offer));
-
-        // Create and register listeners for provider and consumer
-
-        providerObservable.registerListener(negotiationListenerMock);
-        consumerObservable.registerListener(negotiationListenerMock);
 
         // Start provider and consumer negotiation managers
         providerManager.start();
         consumerManager.start();
 
         // Create an initial request and trigger consumer manager
-        ContractOfferRequest request = ContractOfferRequest.Builder.newInstance()
+        var request = ContractOfferRequest.Builder.newInstance()
                 .connectorId("connectorId")
                 .connectorAddress("connectorAddress")
                 .contractOffer(offer)
@@ -136,24 +211,19 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
                     assertThat(providerNegotiation.getContractAgreement()).isNull();
                     verify(validationService, atLeastOnce()).validate(token, offer);
                 });
-
-        // Stop provider and consumer negotiation managers
-        providerManager.stop();
-        consumerManager.stop();
     }
 
     @Test
     void testNegotiation_agreementDeclined() {
+        when(providerDispatcherRegistry.send(any(), isA(ContractAgreementRequest.class), any())).then(onProviderSentAgreementRequest());
+        when(consumerDispatcherRegistry.send(any(), isA(ContractOfferRequest.class), any())).then(onConsumerSentOfferRequest());
+        when(consumerDispatcherRegistry.send(any(), isA(ContractRejection.class), any())).then(onConsumerSentRejection());
         consumerNegotiationId = null;
         var offer = getContractOffer();
 
         when(validationService.validate(token, offer)).thenReturn(Result.success(offer));
         when(validationService.validate(eq(token), any(ContractAgreement.class),
                 any(ContractOffer.class))).thenReturn(false);
-
-        // Create and register listeners for provider and consumer
-        providerObservable.registerListener(negotiationListenerMock);
-        consumerObservable.registerListener(negotiationListenerMock);
 
         // Start provider and consumer negotiation managers
         providerManager.start();
@@ -186,10 +256,6 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
                     verify(validationService, atLeastOnce()).validate(token, offer);
                     verify(validationService, atLeastOnce()).validate(eq(token), any(ContractAgreement.class), any(ContractOffer.class));
                 });
-
-        // Stop provider and consumer negotiation managers
-        providerManager.stop();
-        consumerManager.stop();
     }
 
     @Test
@@ -204,16 +270,12 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
         when(validationService.validate(eq(token), any(ContractAgreement.class),
                 eq(counterOffer))).thenReturn(true);
 
-        // Create and register listeners for provider and consumer
-        providerObservable.registerListener(negotiationListenerMock);
-        consumerObservable.registerListener(negotiationListenerMock);
-
         // Start provider and consumer negotiation managers
         providerManager.start();
         consumerManager.start();
 
         // Create an initial request and trigger consumer manager
-        ContractOfferRequest request = ContractOfferRequest.Builder.newInstance()
+        var request = ContractOfferRequest.Builder.newInstance()
                 .connectorId("connectorId")
                 .connectorAddress("connectorAddress")
                 .contractOffer(initialOffer)
@@ -245,9 +307,6 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
                     verify(validationService, atLeastOnce()).validate(token, counterOffer, initialOffer);
                     verify(validationService, atLeastOnce()).validate(eq(token), any(ContractAgreement.class), any(ContractOffer.class));
                 });
-        // Stop provider and consumer negotiation managers
-        providerManager.stop();
-        consumerManager.stop();
     }
 
     @Test
@@ -261,16 +320,12 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
         when(validationService.validate(token, initialOffer)).thenReturn(Result.success(null));
         when(validationService.validate(token, counterOffer, initialOffer)).thenReturn(Result.success(null));
 
-        // Create and register listeners for provider and consumer
-        providerObservable.registerListener(negotiationListenerMock);
-        consumerObservable.registerListener(negotiationListenerMock);
-
         // Start provider and consumer negotiation managers
         providerManager.start();
         consumerManager.start();
 
         // Create an initial request and trigger consumer manager
-        ContractOfferRequest request = ContractOfferRequest.Builder.newInstance()
+        var request = ContractOfferRequest.Builder.newInstance()
                 .connectorId("connectorId")
                 .connectorAddress("connectorAddress")
                 .contractOffer(initialOffer)
@@ -302,9 +357,6 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
                     verify(validationService, atLeastOnce()).validate(token, counterOffer, initialOffer);
                     verify(validationService, atLeastOnce()).validate(eq(token), any(ContractAgreement.class), any(ContractOffer.class));
                 });
-        // Stop provider and consumer negotiation managers
-        providerManager.stop();
-        consumerManager.stop();
     }
 
     @Test
@@ -329,16 +381,13 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
         // Mock validation of agreement on consumer side
         when(validationService.validate(eq(token), any(ContractAgreement.class),
                 eq(consumerCounterOffer))).thenReturn(true);
-        // Create and register listeners for provider and consumer
-        providerObservable.registerListener(negotiationListenerMock);
-        consumerObservable.registerListener(negotiationListenerMock);
 
         // Start provider and consumer negotiation managers
         providerManager.start();
         consumerManager.start();
 
         // Create an initial request and trigger consumer manager
-        ContractOfferRequest request = ContractOfferRequest.Builder.newInstance()
+        var request = ContractOfferRequest.Builder.newInstance()
                 .connectorId("connectorId")
                 .connectorAddress("connectorAddress")
                 .contractOffer(initialOffer)
@@ -375,9 +424,6 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
                     verify(validationService, atLeastOnce()).validate(token, consumerCounterOffer, counterOffer);
                     verify(validationService, atLeastOnce()).validate(eq(token), any(ContractAgreement.class), eq(consumerCounterOffer));
                 });
-        // Stop provider and consumer negotiation managers
-        providerManager.stop();
-        consumerManager.stop();
     }
 
     @Test
@@ -399,16 +445,12 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
         //Mock validation of second counter offer on provider side => decline
         when(validationService.validate(token, consumerCounterOffer, counterOffer)).thenReturn(Result.success(null));
 
-        // Create and register listeners for provider and consumer
-        providerObservable.registerListener(negotiationListenerMock);
-        consumerObservable.registerListener(negotiationListenerMock);
-
         // Start provider and consumer negotiation managers
         providerManager.start();
         consumerManager.start();
 
         // Create an initial request and trigger consumer manager
-        ContractOfferRequest request = ContractOfferRequest.Builder.newInstance()
+        var request = ContractOfferRequest.Builder.newInstance()
                 .connectorId("connectorId")
                 .connectorAddress("connectorAddress")
                 .contractOffer(initialOffer)
@@ -445,9 +487,45 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
                     verify(validationService, atLeastOnce()).validate(token, counterOffer, initialOffer);
                     verify(validationService, atLeastOnce()).validate(token, consumerCounterOffer, counterOffer);
                 });
-        // Stop provider and consumer negotiation managers
-        providerManager.stop();
-        consumerManager.stop();
+    }
+
+    private Answer<Object> onConsumerSentOfferRequest() {
+        return i -> {
+            ContractOfferRequest request = i.getArgument(1);
+            consumerNegotiationId = request.getCorrelationId();
+            var result = providerManager.offerReceived(token, request.getCorrelationId(), request.getContractOffer(), "hash");
+            if (result.fatalError()) {
+                result = providerManager.requested(token, request);
+            }
+            return toFuture(result);
+        };
+    }
+
+    @NotNull
+    private Answer<Object> onConsumerSentRejection() {
+        return i -> {
+            ContractRejection request = i.getArgument(1);
+            var result = providerManager.declined(token, request.getCorrelationId());
+            return toFuture(result);
+        };
+    }
+
+    @NotNull
+    private Answer<Object> onProviderSentAgreementRequest() {
+        return i -> {
+            ContractAgreementRequest request = i.getArgument(1);
+            var result = consumerManager.confirmed(token, request.getCorrelationId(), request.getContractAgreement(), request.getPolicy());
+            return toFuture(result);
+        };
+    }
+
+    @NotNull
+    private static CompletableFuture<?> toFuture(StatusResult<ContractNegotiation> result) {
+        if (result.succeeded()) {
+            return completedFuture("Success!");
+        } else {
+            return failedFuture(new Exception("Negotiation failed."));
+        }
     }
 
     private void assertNegotiations(ContractNegotiation consumerNegotiation, ContractNegotiation providerNegotiation, int expectedSize) {
@@ -455,6 +533,82 @@ class ContractNegotiationIntegrationTest extends AbstractContractNegotiationInte
         assertThat(providerNegotiation).isNotNull();
         assertThat(consumerNegotiation.getContractOffers()).hasSize(expectedSize);
         assertThat(consumerNegotiation.getContractOffers().size()).isEqualTo(providerNegotiation.getContractOffers().size());
+    }
+
+    /**
+     * Creates the initial contract offer.
+     *
+     * @return the contract offer.
+     */
+    private ContractOffer getContractOffer() {
+        return ContractOffer.Builder.newInstance()
+                .id(ContractId.createContractId("1"))
+                .contractStart(ZonedDateTime.now())
+                .contractEnd(ZonedDateTime.now().plusMonths(1))
+                .provider(URI.create("provider"))
+                .consumer(URI.create("consumer"))
+                .asset(Asset.Builder.newInstance().build())
+                .policy(Policy.Builder.newInstance()
+                        .type(PolicyType.CONTRACT)
+                        .assigner("assigner")
+                        .assignee("assignee")
+                        .duty(Duty.Builder.newInstance()
+                                .action(Action.Builder.newInstance()
+                                        .type("USE")
+                                        .build())
+                                .build())
+                        .build())
+                .build();
+    }
+
+    /**
+     * Creates the first counter offer.
+     *
+     * @return the contract offer.
+     */
+    private ContractOffer getCounterOffer() {
+        return ContractOffer.Builder.newInstance()
+                .id(UUID.randomUUID().toString())
+                .contractStart(ZonedDateTime.now())
+                .contractEnd(ZonedDateTime.now().plusMonths(2))
+                .provider(URI.create("provider"))
+                .consumer(URI.create("consumer"))
+                .policy(Policy.Builder.newInstance()
+                        .type(PolicyType.CONTRACT)
+                        .assigner("assigner")
+                        .assignee("assignee")
+                        .duty(Duty.Builder.newInstance()
+                                .action(Action.Builder.newInstance()
+                                        .type("USE")
+                                        .build())
+                                .build())
+                        .build())
+                .build();
+    }
+
+    /**
+     * Creates the second counter offer.
+     *
+     * @return the contract offer.
+     */
+    private ContractOffer getConsumerCounterOffer() {
+        return ContractOffer.Builder.newInstance()
+                .id(UUID.randomUUID().toString())
+                .contractStart(ZonedDateTime.now())
+                .contractEnd(ZonedDateTime.now().plusMonths(3))
+                .provider(URI.create("provider"))
+                .consumer(URI.create("consumer"))
+                .policy(Policy.Builder.newInstance()
+                        .type(PolicyType.CONTRACT)
+                        .assigner("assigner")
+                        .assignee("assignee")
+                        .duty(Duty.Builder.newInstance()
+                                .action(Action.Builder.newInstance()
+                                        .type("USE")
+                                        .build())
+                                .build())
+                        .build())
+                .build();
     }
 
 }
