--- core/federated-catalog/federated-catalog-core/src/test/java/org/eclipse/dataspaceconnector/catalog/cache/FccTestExtension.java
@@ -1,191 +1,0 @@
-/*
- *  Copyright (c) 2020 - 2022 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *
- */
-
-package org.eclipse.dataspaceconnector.catalog.cache;
-
-import kotlin.NotImplementedError;
-import org.eclipse.dataspaceconnector.catalog.directory.InMemoryNodeDirectory;
-import org.eclipse.dataspaceconnector.catalog.spi.FederatedCacheNodeDirectory;
-import org.eclipse.dataspaceconnector.policy.model.Action;
-import org.eclipse.dataspaceconnector.policy.model.Permission;
-import org.eclipse.dataspaceconnector.policy.model.Policy;
-import org.eclipse.dataspaceconnector.policy.model.PolicyType;
-import org.eclipse.dataspaceconnector.runtime.metamodel.annotation.Provides;
-import org.eclipse.dataspaceconnector.spi.contract.negotiation.ContractNegotiationManager;
-import org.eclipse.dataspaceconnector.spi.contract.offer.ContractOfferQuery;
-import org.eclipse.dataspaceconnector.spi.contract.offer.ContractOfferService;
-import org.eclipse.dataspaceconnector.spi.contract.offer.store.ContractDefinitionStore;
-import org.eclipse.dataspaceconnector.spi.iam.IdentityService;
-import org.eclipse.dataspaceconnector.spi.message.MessageContext;
-import org.eclipse.dataspaceconnector.spi.message.Range;
-import org.eclipse.dataspaceconnector.spi.message.RemoteMessageDispatcher;
-import org.eclipse.dataspaceconnector.spi.message.RemoteMessageDispatcherRegistry;
-import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
-import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
-import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
-import org.eclipse.dataspaceconnector.spi.transfer.store.TransferProcessStore;
-import org.eclipse.dataspaceconnector.spi.types.domain.asset.Asset;
-import org.eclipse.dataspaceconnector.spi.types.domain.contract.offer.ContractDefinition;
-import org.eclipse.dataspaceconnector.spi.types.domain.contract.offer.ContractOffer;
-import org.eclipse.dataspaceconnector.spi.types.domain.message.RemoteMessage;
-import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcess;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import java.util.stream.Stream;
-
-import static java.util.Collections.emptyList;
-
-@Provides({ RemoteMessageDispatcherRegistry.class, TransferProcessStore.class, ContractDefinitionStore.class, IdentityService.class, ContractNegotiationManager.class, FederatedCacheNodeDirectory.class })
-public class FccTestExtension implements ServiceExtension {
-
-    @Override
-    public void initialize(ServiceExtensionContext context) {
-        List<Asset> assets = Collections.emptyList();
-        context.registerService(TransferProcessStore.class, new FakeTransferProcessStore());
-        context.registerService(RemoteMessageDispatcherRegistry.class, new FakeRemoteMessageDispatcherRegistry());
-        context.registerService(ContractOfferService.class, new FakeContractOfferService(assets));
-        context.registerService(ContractDefinitionStore.class, new FakeContractDefinitionStore());
-        context.registerService(FederatedCacheNodeDirectory.class, new InMemoryNodeDirectory());
-    }
-
-    private static class FakeContractOfferService implements ContractOfferService {
-        private final List<Asset> assets;
-
-        private FakeContractOfferService(List<Asset> assets) {
-            this.assets = assets;
-        }
-
-        @Override
-        @NotNull
-        public Stream<ContractOffer> queryContractOffers(ContractOfferQuery query, Range range) {
-            return assets.stream().map(asset ->
-                    ContractOffer.Builder.newInstance()
-                            .id("1")
-                            .policy(createEverythingAllowedPolicy())
-                            .asset(asset)
-                            .build()
-            );
-        }
-
-        private Policy createEverythingAllowedPolicy() {
-            var policyBuilder = Policy.Builder.newInstance();
-            var permissionBuilder = Permission.Builder.newInstance();
-            var actionBuilder = Action.Builder.newInstance();
-
-            policyBuilder.type(PolicyType.CONTRACT);
-            actionBuilder.type("USE");
-            permissionBuilder.target("1");
-
-            permissionBuilder.action(actionBuilder.build());
-            policyBuilder.permission(permissionBuilder.build());
-
-            policyBuilder.target("1");
-            return policyBuilder.build();
-        }
-    }
-
-    private static class FakeTransferProcessStore implements TransferProcessStore {
-        @Override
-        public TransferProcess find(String id) {
-            return null;
-        }
-
-        @Override
-        public @Nullable String processIdForDataRequestId(String id) {
-            return null;
-        }
-
-        @Override
-        public void create(TransferProcess process) {
-        }
-
-        @Override
-        public void update(TransferProcess process) {
-        }
-
-        @Override
-        public void delete(String processId) {
-        }
-
-        @Override
-        public Stream<TransferProcess> findAll(QuerySpec querySpec) {
-            return null;
-        }
-
-        @Override
-        public @NotNull List<TransferProcess> nextForState(int state, int max) {
-            return emptyList();
-        }
-    }
-
-    private static class FakeRemoteMessageDispatcherRegistry implements RemoteMessageDispatcherRegistry {
-
-        @Override
-        public void register(RemoteMessageDispatcher dispatcher) {
-        }
-
-        @Override
-        public <T> CompletableFuture<T> send(Class<T> responseType, RemoteMessage message, MessageContext context) {
-            return null;
-        }
-    }
-
-    private static class FakeContractDefinitionStore implements ContractDefinitionStore {
-
-        private final List<ContractDefinition> contractDefinitions = new ArrayList<>();
-
-        @Override
-        public @NotNull Stream<ContractDefinition> findAll(QuerySpec spec) {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public ContractDefinition findById(String definitionId) {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public void save(Collection<ContractDefinition> definitions) {
-            contractDefinitions.addAll(definitions);
-        }
-
-        @Override
-        public void save(ContractDefinition definition) {
-            contractDefinitions.add(definition);
-        }
-
-        @Override
-        public void update(ContractDefinition definition) {
-            throw new NotImplementedError();
-        }
-
-        @Override
-        public ContractDefinition deleteById(String id) {
-            throw new NotImplementedError();
-        }
-
-        @Override
-        public void reload() {
-            throw new NotImplementedError();
-        }
-    }
-
-
-}
